bank/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.3
Django==5.2.3
django-allauth==65.13.0
django-celery-beat==2.8.1
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-defender==0.9.8
django-json-widget==2.0.1
django-simpleui==2025.5.262
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.2.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.9.0
